src/Elksoft.Converters/Elksoft.Converters.UncheckedNumericConverters/UncheckedUIntPtrToIntPtrConverter.cs
// <copyright file="UncheckedUIntPtrToIntPtrConverter.cs" company="Georgy Losenkov">
// Copyright (c) Georgy Losenkov. All rights reserved.
// </copyright>

#if NET7_0_OR_GREATER
using System;

namespace Elksoft.Converters.UncheckedNumericConverters
{
    internal sealed class UncheckedUIntPtrToIntPtrConverter : Converter<UIntPtr, IntPtr>
    {
        public override Boolean AcceptsNull => false;

        public override Boolean IsExplicit => true;

        public override IntPtr Convert(UIntPtr value, IFormatProvider? formatProvider)
        {
            return unchecked((IntPtr)value);
        }
    }
}
#endif
